requirements.txt
Flask==2.0.1
requests==2.28.2
gunicorn==20.1.0
python-dotenv==0.21.1 